refactor: Phase 1 tech debt — PHPStan zero baseline, 93 tests, exception cleanup, pagination
- PHPStan baseline 35→0: fix type annotations, dead catches, logic issues, Collection covariance
- 93 new tests: job tests (GenerateReport, RunSafeUpdate), 8 report gatherer tests, 3 storage driver tests (Local, S3, Dropbox)
- Exception cleanup: generic catches → QueryException, RuntimeException, ExternalServiceException
- Pagination: StatusPagesList, ReportTemplatesSettings, PresetManager
- CI: PHPStan baseline guard (max 0), test coverage reporting with pcov
- Bug fixes: UptimeGatherer avg() float cast, getenv()→config(), glob() safe return, openssl pass arg

+++ app/Services/Reports/Sections/UptimeGatherer.php
@@ -50,7 +50,7 @@ class UptimeGatherer extends BaseReportSectionGatherer
             ->get()
             ->toArray();
 
-        $avgResponseTime = UptimeCheck::where('monitor_id', $monitor->id)
+        $avgResponseTime = (float) UptimeCheck::where('monitor_id', $monitor->id)
             ->whereBetween('checked_at', [$periodStart, $periodEnd])
             ->where('is_up', true)
             ->avg('response_time');
